% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{accuracy_matrix}
\alias{accuracy_matrix}
\title{Generates matrix of accuracy values}
\usage{
accuracy_matrix(rf, svm)
}
\arguments{
\item{rf}{A random forest object}

\item{svm}{A support vector machine object}
}
\description{
Returns a matrix of accuracy values where the columns represent the classifier type, and the rows represent the min_peak_count.neighbor values used for the classifier
}

