% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{combine_peaks}
\alias{combine_peaks}
\title{Combine peak vectors}
\usage{
combine_peaks(list_mz_peaks)
}
\arguments{
\item{list_mz_peaks}{List of m/z peak vectors}
}
\description{
Create a binary matrix, each column represents an m/z value, and each row represents a mass spectra.  The value indicates whether or not the m/z of this spectra is a peak.
}

