% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persondf.R
\name{dfmerger}
\alias{dfmerger}
\title{Reduce to one variable helper}
\usage{
dfmerger(filldf, srcdf)
}
\arguments{
\item{filldf}{}
}
\description{
Takes in a data frame whose freq column hasn't been modified and fills it with matching rows from srcdf
}
\examples{
dfmerger()
}

