% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{forestSelect}
\alias{forestSelect}
\title{forestSelect}
\usage{
forestSelect(m, lbl, r)
}
\arguments{
\item{nxd}{matrix with n individuals and d dimensions}

\item{vector}{of length n with integer labels}

\item{number}{of features to select (r<d)}
}
\description{
Get top r features using data matrix and labels for data
}
\examples{
forestSelect()
}

