% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{getPrComps}
\alias{getPrComps}
\title{getPrComps}
\usage{
getPrComps(spectrums)
}
\arguments{
\item{spectrums}{nxd matrix with n vectors of d dimensions}
}
\description{
Get principal components that explain 90% of variance.
}
\examples{
getPrComps()
}

