% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{kmeansCH}
\alias{kmeansCH}
\title{kmeans CH}
\usage{
kmeansCH(f)
}
\arguments{
\item{nxd}{matrix with n individuals and d dimensions}
}
\description{
Run k-means, using the k resulting in best CH value
}
\examples{
kmeansCH()
}

