% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{kmeansPCA}
\alias{kmeansPCA}
\title{kmeansPCA}
\usage{
kmeansPCA(spectrums)
}
\arguments{
\item{spectrums}{nxm matrix representing frequencies for n individuals over m integer m/z values.}
}
\description{
Run getPrComps and k-means
}
\examples{
kmeansPCA()
}

