% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persondf.R
\name{minMaxByDay}
\alias{minMaxByDay}
\title{Returns the min and max of each day}
\usage{
minMaxByDay(...)
}
\arguments{
\item{...}{Each arameter should be a person df}
}
\description{
Reduces each persons m/z, frequencies, and days to just frequencies
}
\examples{
minMaxByDay()
}

