% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{naive_feature_importance}
\alias{naive_feature_importance}
\title{Rank importance of features}
\usage{
naive_feature_importance(peaks, labels)
}
\arguments{
\item{peaks}{A matrix of peaks}

\item{labels}{A factor vector of labels whose length is equal to the number of rows of peaks}
}
\description{
Given a matrix of binary peaks and each row's corresponding labels, this function takes returns the absolute difference between the proportion of times an m/z value was labeled as a peak within each of the two classes.
}

