% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persondf.R
\name{persondf.mean}
\alias{persondf.mean}
\title{Mean frequencies parameters}
\usage{
persondf.mean(ids, bigdf)
}
\arguments{
\item{ids}{Identities to be combined}

\item{bigdf}{Data frame of rbinded persondfs}
}
\description{
Mean frequencies parameters
}
\examples{
persondf.mean()
}

