% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persondf.R
\name{persondf.plot}
\alias{persondf.plot}
\title{Plot person}
\usage{
persondf.plot(persondf)
}
\arguments{
\item{persondf}{Person data frame}
}
\description{
Plots the person by day
}
\examples{
persondf.plot()
}

