% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persondf.R
\name{persondf}
\alias{persondf}
\title{Person constructor}
\usage{
persondf(id, dates, ...)
}
\arguments{
\item{id}{What is this person's name}

\item{...}{All the data frames of m/z values and frequencies}

\item{days}{Vector of days since initial sample taken}
}
\description{
Creates a data frame with ids, m/z values, frequencies, and days.
}
\examples{
persondf()
}

