% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{plot_naive_importance}
\alias{plot_naive_importance}
\title{Plot importance of naive importance vector}
\usage{
plot_naive_importance(naive_importance)
}
\arguments{
\item{naive_importance}{A vector returned from naive_feature_importance()}
}
\description{
Plot importance of each m/z value according to the naive ranking vector.  Returns a ggplot object
}

