% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{plot_rf_differences}
\alias{plot_rf_differences}
\title{Plot important peaks on each day}
\usage{
plot_rf_differences(peak_freqs, rf, count)
}
\arguments{
\item{rf}{A random forest object, to be used for feature importance}

\item{count}{Number of peaks to select}

\item{peaks_freqs}{The frequency of each peak within each label}
}
\description{
Returns a ggplot object of the most important peak frequencies
}

