% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{plot_rf_importance}
\alias{plot_rf_importance}
\title{Plot importance of randomForest}
\usage{
plot_rf_importance(rf, count)
}
\arguments{
\item{rf}{A randomForest object}
}
\description{
Plot importance of each m/z value according to a randomForest object.  Returns a ggplot object
}

