% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{plotRFaccuracy}
\alias{plotRFaccuracy}
\title{Plot importance of randomForest}
\usage{
plotRFaccuracy(rf)
}
\arguments{
\item{rf}{A randomForest object}
}
\description{
Plot importance of each m/z value according to a randomForest object.  Returns a ggplot object
}

