% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{randIndex}
\alias{randIndex}
\title{randIndex}
\usage{
randIndex(oc, nc)
}
\arguments{
\item{vector}{of length n with integer labels}

\item{vector}{of length n with integer labels}
}
\description{
Measure similarity of two clustering assignments
}
\examples{
randIndex()
}

