% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{random_forest_classify}
\alias{random_forest_classify}
\title{Random Forest}
\usage{
random_forest_classify(training_set, training_labels)
}
\description{
Build a random forest classifier and test it.  No need for test set because out-of-bag error measurement.
}

