% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persondf.R
\name{reduceAndConcat}
\alias{reduceAndConcat}
\title{Reduce to one variable}
\usage{
reduceAndConcat(...)
}
\arguments{
\item{...}{Each parameter should be a person df}
}
\description{
Reduces each persons m/z, frequencies, and days to just frequencies
}
\examples{
reduceAndConcat()
}

