% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{binary_peaks}
\alias{binary_peaks}
\title{Find binary peaks}
\usage{
binary_peaks(df, neighbors, error = 0)
}
\arguments{
\item{df}{Data frame of m/z and intensities}

\item{neighbors}{Number of neighboring m/z values to compare on right and left}

\item{error}{m/z measuring error parameter}
}
\description{
Find binary peaks
}
\examples{
binary_peaks()
}

