% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{classifier_accuracies}
\alias{classifier_accuracies}
\title{Classifier Accuracies}
\usage{
classifier_accuracies(peaks, labels, min_peak_percentage)
}
\arguments{
\item{peaks}{Boolean matrix of mass spectra rows with m/z columns, indicating if an m/z value corresponds to a peak.}

\item{labels}{The correct classifications of the peaks.}

\item{minpeaks}{How many "true" values must show up for a given m/z value for it to be considered a feature.}
}
\description{
Find the best classifier using leave-one-out cross validation (svm) and out-of-bag error (random forests).  Returns a list of classifier results
}

