% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{getPrComps}
\alias{getPrComps}
\title{getPrComps}
\usage{
getPrComps(spectrums)
}
\arguments{
\item{spectrums}{nxm matrix representing frequencies for n individuals over m integer m/z values.}
}
\description{
Get principal components that explain 90% of variance.
}
\examples{
getPrComps()
}

