% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/persondf.R
\name{persondf.mean}
\alias{persondf.mean}
\title{Mean frequencies parameters}
\usage{
persondf.mean(...)
}
\arguments{
\item{...}{Each parameter should be a persondf object}
}
\description{
Mean frequencies parameters
}
\examples{
persondf.mean()
}

